Extract deletion handlers
